porcupine/requirements.txt
typer[all]==0.12.3
rich==13.7.1
litellm==1.44.22
py-clob-client==0.20.0
supabase==2.7.4
keyring==25.2.1
python-dotenv==1.0.1
httpx==0.27.0
tenacity==8.3.0
